Allow empty au-sort attribute
This change allows table columns to be generated using an iterator. For example (using PUG notation):
```th(repeat.for='col of cols' aut-sort='key.bind: col.sortField')```
can leave it up to individual columns to decide whether there is a key to be bound. 
The workaround is to declare two ```th``` attributes, one with ```if.bind='col.sortField'``` and the other without.